First Draft
This is just the first draft so we can start building this feature.

- Added dataloader.py, which loads data for training
- Added train.py, with the current training loop
- Added lora.py, for LoRA wrapper of the stage 1 Transformer
- Added dummy_dataset folder with 25 data samples to work with when testing (VCTK-->p311)
- Commented out the initial inference code when stage 1 model is built.

There is no batch processing in the training loop currently (was getting some dimension mismatching in the KVCache.update).

The dataloader works fine, but everything else requires some work. This is just an initial draft so we can start working on this thing together! :-)