move loss to model + format similar to nanoGPT
Moved loss calculation to LoRA wrapper model.

Modified training loop to be similar to that of nanoGPT. This involves using a sliding window for prompts & labels, which should more accurately replicate what the model is actually producing at logits level. If my intuition is wrong about this, please correct me.